fix: addressing feedback, folder rename, exporting missing types
