+++ src/event-bridge/event-pattern/constants.ts
