+++ src/event-bridge/event-pattern/numeric.ts
