+++ src/event-bridge/event-pattern/pattern.ts
