+++ src/event-bridge/event-pattern/synth.ts
