+++ src/event-bridge/event-pattern/utils.ts
