+++ src/event-bridge/index.ts
@@ -1,0 +1,9 @@
+export * from "./bus";
+export * from "./rule";
+export * from "./types";
+// Do not export `pipe`
+export {
+  EventBusTargetProps,
+  EventBusTargetResource,
+  LambdaTargetProps,
+} from "./target";
