+++ src/event-bridge/target-input/index.ts
