--- src/eventbridge/index.ts
@@ -1,3 +1,0 @@
-export * from "./bus";
-export * from "./rule";
-export * from "./types";
